Rename main package and class
Rename seedu.duke package to expiryeliminator
Rename Duke class to ExpiryEliminator
