+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Duke"
+    mainClassName = "expiryeliminator.ExpiryEliminator"
 }
 
 shadowJar {
