+++ src/main/java/expiryeliminator/ExpiryEliminator.java
@@ -1,8 +1,8 @@
-package seedu.duke;
+package expiryeliminator;
 
 import java.util.Scanner;
 
-public class Duke {
+public class ExpiryEliminator {
     /**
      * Main entry-point for the java.duke.Duke application.
      */
