+++ src/test/java/expiryeliminator/ExpiryEliminatorTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package expiryeliminator;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class ExpiryEliminatorTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
